[feature #148963113] rename files in PascalCase
