+++ server/test/server/User_spec.js
@@ -1,5 +1,5 @@
 /* global describe it */
-const app = require('./../../../app');
+const app = require('./../../../App');
 const request = require('supertest');
 const db = require('./../../models');
